--- a/blendsql/__init__.py
+++ b/blendsql/__init__.py
@@ -1,5 +1,2 @@
-__version__ = "0.0.21"
-
-
 from .ingredients.builtin import LLMMap, LLMQA, LLMJoin, LLMValidate, ImageCaption
 from .blend import blend