--- a/blendsql/_constants.py
+++ b/blendsql/_constants.py
@@ -11,7 +11,7 @@ class StrInMeta(EnumMeta):
 
 DEFAULT_ANS_SEP = ";"
 DEFAULT_NAN_ANS = "-"
-MAP_BATCH_SIZE = 5
+MAP_BATCH_SIZE = 15
 
 
 class IngredientType(str, Enum, metaclass=StrInMeta):