--- a/blendsql/generate/__init__.py
+++ b/blendsql/generate/__init__.py
@@ -1,3 +1,0 @@
-from .text import text
-from .regex import regex
-from .choice import choice