--- a/tox.ini
+++ b/tox.ini
@@ -13,6 +13,7 @@ deps =
     duckdb
     ollama
     pillow
+    accelerate
 
 [testenv:coverage]
 basepython = python3.12