This PR moves away from outlines and back to guidance for the underlying constrained decoding. This improves runtime of executing BlendSQL queries, and allows us to execute ingredients with greater guarantees on the output format by interleaving text + generations. 

Additional changes:
- Moved to pyproject.toml format (away from setuptools.py)
- Added `AnthropicLLM`, removed `LlamaCpp`

From now on, I use the `HuggingFaceTB/SmolLM-135M` model for benchmarks. 

Before:
| Task           |   Average Runtime |   # Unique Queries |
|:---------------|------------------:|-------------------:|
| financials     |         0.0427749 |                  7 |
| rugby          |         3.54232   |                  4 |
| national_parks |         2.63405   |                  5 |
| 1966_nba_draft |         3.65771   |                  2 |

After:
| Task           |   Average Runtime |   # Unique Queries |
|:---------------|------------------:|-------------------:|
| financials     |         0.0487881 |                  7 |
| rugby          |         0.909974  |                  4 |
| national_parks |         2.13209   |                  5 |
| 1966_nba_draft |         1.39948   |                  2 |